Corrections

+++ backend/controllers/sauces.js
@@ -1,6 +1,7 @@
 const Sauce = require('../models/Sauce');
 const fs = require('fs');
 const jwt = require("jsonwebtoken");
+const { Error } = require('mongoose');
 
 
 exports.createSauce = (req, res, next) => {
@@ -11,31 +12,57 @@ exports.createSauce = (req, res, next) => {
         imageUrl: `${req.protocol}://${req.get('host')}/images/${req.file.filename}`
     });
     sauce.save()
-        .then(() => res.status(201).json({ message: 'Sauce enregistré !' }))
+        .then(() => res.status(201).json({ message: 'Sauce enregistrée !' }))
         .catch(error => res.status(400).json({ error }));
 };
 
 
 exports.modifySauce = (req, res, next) => {
-    const sauceObject = req.file ?
-        {
-            ...JSON.parse(req.body.sauce),
-            imageUrl: `${req.protocol}://${req.get('host')}/images/${req.file.filename}`
-        } : { ...req.body };
-    Sauce.updateOne({ _id: req.params.id }, { ...sauceObject, _id: req.params.id })
-        .then(() => res.status(201).json({ message: 'Sauce modifiée !' }))
-        .catch(error => res.status(400).json({ error }));
+    Sauce.findOne({ _id: req.params.id })
+        .then((sauce) => {
+            if (!sauce) {
+                res.status(404).json({
+                    message: 'Sauce non trouvée !'
+                });
+            }
+            else if (sauce.userId !== req.auth.userId) {
+                res.status(401).json({
+                    message: 'Requête non autorisée !'
+                });
+            } else {
+                const sauceObject = req.file ?
+                    {
+                        ...JSON.parse(req.body.sauce),
+                        imageUrl: `${req.protocol}://${req.get('host')}/images/${req.file.filename}`
+                    } : { ...req.body };
+                Sauce.updateOne({ _id: req.params.id }, { ...sauceObject, _id: req.params.id })
+                    .then(() => res.status(201).json({ message: 'Sauce modifiée !' }))
+                    .catch(error => res.status(400).json({ error }));
+            }
+        })
 };
 
 exports.deleteSauce = (req, res, next) => {
     Sauce.findOne({ _id: req.params.id })
-        .then(sauce => {
-            const filename = sauce.imageUrl.split('/images/')[1];
-            fs.unlink(`images/${filename}`, () => {
-                Sauce.deleteOne({ _id: req.params.id })
-                    .then(() => res.status(200).json({ message: 'Objet supprimé !' }))
-                    .catch(error => res.status(400).json({ error }));
-            });
+        .then((sauce) => {
+            if (!sauce) {
+                res.status(404).json({
+                    message: 'Sauce non trouvée !'
+                });
+            }
+            else if (sauce.userId !== req.auth.userId) {
+                res.status(401).json({
+                    message: 'Requête non autorisée !'
+                });
+            } else {
+                const filename = sauce.imageUrl.split('/images/')[1];
+                fs.unlink(`images/${filename}`, () => {
+                    Sauce.deleteOne({ _id: req.params.id })
+                        .then(() => res.status(200).json({ message: 'Sauce supprimée !' }))
+                        .catch(error => res.status(400).json({ error }));
+                });
+            }
+
         })
         .catch(error => res.status(500).json({ error }));
 
@@ -72,21 +99,12 @@ exports.likeSauce = (req, res, next) => {
             {
                 $inc: { dislikes: req.body.like++ * -1 },
                 $push: { usersDisliked: req.body.userId },
+
             }
         )
             .then(sauce => res.status(200).json({ message: "Nouveau dislike !" }))
             .catch(error => res.status(400).json({ error }));
     }
-
-
-
-
-
-
-
-
-
-
     else {
         Sauce.findOne({ _id: req.params.id })
             .then((sauce) => {
